Refactor new code into repo.go
